requirements.txt
chromadb
langchain
langchain_community
langchain_core
langchain_openai
openai
python-dotenv
streamlit
unstructured
unstructured[md]
langchain-text-splitters
